# openapi.yaml
openapi: 3.1.0
info:
  title: komondor-api
  version: "1.0.0"
  summary: TSL Sequence Data API — projects, samples, sequencing runs and their files.
  description: |
    A description of what komondor-api **does today**, derived by reading
    `routes/` and `models/`. It is a record of current behaviour, not a
    proposal. Where the current behaviour is a known defect it is documented as
    it is, and the operation carries a `DRIFT` or `DEFECT` note saying so.
    Fixing one of those is a change to the API and must update this file in the
    same commit.

    `__tests__/contract/openapi.test.js` fails if a route exists in `routes/`
    with no entry here, so the two cannot silently diverge.

    ## Known drift from the sibling clients

    Three repositories consume this API: `komondor-web` (Nuxt 2 UI),
    `komondor-power` (bulk CSV ingest) and `komondor-nudge` (a scheduled
    worker that reads the same MongoDB directly rather than over HTTP). The
    following are the places where their assumptions and this server's
    behaviour have come apart. See `docs/CONTRACTS.md` for the policy.

    1. **Routes are mounted at the ROOT, not under `/api`.** `app.js` calls
       `app.use(authRoutes)` with no path prefix, so login is `POST /login`,
       not `POST /api/login` or `POST /auth/login`. The comment above the login
       handler in `routes/auth.js` still reads "Add POST - /api/login", which
       is where the belief came from. `komondor-power`'s client tries
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /project:
    get:
      tags: [Projects]
      summary: One project, with samples, additional files and on-disk status.
      description: |
        The id is a **query parameter**, not a path segment.

        Alongside the project this returns a reconciliation of the database
        records against the `additional/` directory on disk. A datastore that
        cannot be read degrades to `additionalFilesStatus.status = "UNKNOWN"`
        rather than failing the request. Once a project has left `hpc`, the
        directory is not read: the listing is null and reconciliation is
        `NOT_APPLICABLE`.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: The project and its file reconciliation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                  location:
                    $ref: "#/components/schemas/EntityLocation"
                  actualAdditionalFiles:
                    type: [array, "null"]
                    items:
                      type: string
                    description: Filenames actually present in the project's `additional/` directory.
                  additionalFilesStatus:
                    $ref: "#/components/schemas/FileReconciliation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /samples:
    get:
      tags: [Samples]
      summary: Samples visible to the caller, newest first.
      description: Same visibility rules as `GET /projects`. Sorted in Mongo here.
      responses:
        "200":
          description: Visible samples.
          content:
            application/json:
              schema:
                type: object
                properties:
                  samples:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sample"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/ServerError"
  /samples/names/{projectId}:
    get:
      tags: [Samples]
      summary: Distinct sample names within a project.
      description: |
        Used by clients to pre-empt the idempotent-create branch of
        `POST /samples/new`.

        Requires READ access to the group that owns the project. The project is
        loaded first, so an unknown `projectId` answers 404 and a project in a
        group the caller cannot read answers 403. This endpoint returns every
        sample name in the project, not just the caller's own, so it was
        previously a cross-group read open to any authenticated caller for any
        project id.

        A malformed `projectId` is a 400 and the database is never queried.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: Unique, non-empty names.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sampleNames:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
  /sample:
    get:
      tags: [Samples]
      summary: One sample, with its runs and on-disk file status.
      description: |
        For a Project in `hpc`, reconciles the sample's additional-file
        records with disk. For every other storage state it performs no HPC
        read, returns a null listing and `NOT_APPLICABLE` reconciliation.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: The sample and its file reconciliation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sample:
                    $ref: "#/components/schemas/Sample"
                  location:
                    $ref: "#/components/schemas/EntityLocation"
                  actualAdditionalFiles:
                    type: [array, "null"]
                    items:
                      type: string
                  additionalFilesStatus:
                    $ref: "#/components/schemas/FileReconciliation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
  /samples/new:
    post:
      tags: [Samples]
      summary: Create a sample, or return the existing one.
      description: |
        **IDEMPOTENT CREATE — 200 vs 201.** Before inserting, the handler looks
        for a sample with the same `project` and `name`. If one exists it
        answers **200** with `{ sample, idempotent: true, message }` and the
        PRE-EXISTING document. A genuine insert answers **201** with
        `{ sample }` and no `idempotent` key.

        A client that treats 200 as "created" will double-count, and one that
        treats anything but 201 as failure will break on retry. Branch on the
        status code, or on the presence of `idempotent`.

        `project` is mandatory, so the lookup depends on `name` alone: a
        nameless non-TPlex sample is never deduplicated.

        **The stored `group` is DERIVED from the parent project**, not taken
        from the body. `group` is still required and is cross-checked: a
        `group` that does not own the submitted `project` is a 400 ("The
        submitted group does not own the submitted project."), never a silently
        mis-filed sample. Write access is authorised against the PROJECT's
        group, and that check runs BEFORE the mismatch check so the status code
        cannot be used to probe which group owns a project the caller cannot
        see.

        An unknown `project` answers 404. Malformed ids answer 400 and the
        database is never queried.

        TPlex CSV text is stored with spreadsheet formula characters
        neutralised — a value beginning `=`, `+`, `-`, `@`, tab or CR is
        prefixed with an apostrophe and force-quoted, except plain numbers
        such as `-80`. A consumer round-tripping `tplexCsv` should strip a
        leading `"'` when the field starts with one of those characters.

        TPlex mode: when `tplexCsv` is a non-empty array, ONE sample is created
        with the array flattened to CSV text in `tplexCsv`, the per-organism
        fields forced to null, and a name derived from the first row when none
        was supplied. `additionalFiles` is ignored in this mode.

        If the parent Project has left `hpc`, the authorised request is
        refused before idempotency or file work with 409 and
        `code: PROJECT_STORAGE_READ_ONLY`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group, project, owner]
              properties:
                name:
                  type: string
                project:
                  $ref: "#/components/schemas/ObjectId"
                group:
                  $ref: "#/components/schemas/ObjectId"
                owner:
                  type: string
                scientificName:
                  type: string
                  description: Required unless this is a TPlex sample.
                commonName:
                  type: string
                  description: Required unless this is a TPlex sample.
                ncbi:
                  type: string
                  description: Numeric taxonomy id. Required unless this is a TPlex sample.
                conditions:
                  type: string
                  description: Required unless this is a TPlex sample.
                tplexCsv:
                  type: array
                  description: Rows of a TPlex CSV. Presence of a non-empty array selects TPlex mode.
                  items:
                    type: object
                additionalFiles:
                  type: array
                  items:
                    type: object
      responses:
        "200":
          description: |
            A sample with this name already existed for this project. The body
            is the existing document — nothing was created.
          content:
            application/json:
              schema:
                type: object
                required: [sample, idempotent]
                properties:
                  sample:
                    $ref: "#/components/schemas/Sample"
                  idempotent:
                    const: true
                  message:
                    const: Sample with this name already exists for this project
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sample:
                    $ref: "#/components/schemas/Sample"
        "400":
          description: |
            Missing or malformed `group` / `project`, a `group` that does not
            own the submitted `project`, a TPlex payload whose rows are not
            objects, or mongoose validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          description: |
            The parent `project` does not exist. New: the handler must load the
            project to derive the sample's group.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: The Project is migrating, archived, or has invalid storage state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/ServerError"
  /runs:
    get:
      tags: [Runs]
      summary: Runs visible to the caller, newest first.
      responses:
        "200":
          description: Visible runs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <token>`, signed with `JWT_SECRET` (HS256) by
        `lib/utils/jwtSign.js`.

        **The `exp` claim is mandatory.** `lib/utils/getUserFromRequest.js`
        rejects a structurally valid, correctly signed token that carries no
        `exp` by raising `TokenExpiredError`, which surfaces as 401. This was
        added deliberately (commit "refuse legacy tokens without an expiry
        claim"): tokens issued before August 2026 had no expiry and stayed
        valid forever, including tokens minted during the LDAP incident with an
        empty `groups` claim. Such tokens can only be fixed by re-logging in.

        Tokens expire after `JWT_EXPIRES_IN`, default `7d`.

        **Group membership is a snapshot.** `groups` is resolved once, at
        login, and never re-read. A user added to a group sees nothing new
        until their token expires and they sign in again — that expiry is the
        upper bound on the staleness.

        Requests with NO `Authorization` header are not rejected by the
        app-level middleware; they simply arrive with no `req.user`, and each
        route's `isAuthenticated` guard turns that into a 401. Routes without
        that guard (`/projects/names`, `GET /options/*`, `/health`, `/test`,
        `/login`, `/logout`, `/uploads`) serve anonymous callers.
...
  responses:
    Unauthenticated:
      description: |
        No usable token. Two distinct bodies reach clients as 401:
        `{ error: "Authentication required" }` from a route's
        `isAuthenticated` guard, and
        `{ error: "Invalid or expired authentication token", detail }` from the
        app-level middleware in `app.js` when a token was present but bad or
        expired.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  error:
                    const: Authentication required
              - type: object
                properties:
                  error:
                    const: Invalid or expired authentication token
                  detail:
                    type: string
                    description: The underlying jsonwebtoken message.
                  requestId:
                    type: string
                    description: Present only when raised by the terminal error handler.
...
    GroupPermissionDenied:
      description: |
        The caller is neither the record's owner, nor a member of its group,
        nor an admin. The message embeds the username:
        `User '<username>' does not have permission to ...`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: A required parameter is missing or malformed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFoundEnvelope:
      description: The referenced record does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      # Deliberately not the old hard-coded '2Blades' group id: that literal
      # used to carry meaning in routes/projects.js and reusing it as a generic
      # example invited the reader to think it still does.
      examples: ["507f1f77bcf86cd799439011"]
...
    EntityLocation:
      type: object
      description: Display-ready authoritative location for a detail response.
      properties:
        authoritative:
          type: [string, "null"]
          enum: [hpc, s3, null]
        baseUri:
          type: [string, "null"]
        rawUri:
          type: [string, "null"]
          description: Present on run details only.
        additionalUri:
          type: [string, "null"]
        plannedS3Uri:
          type: [string, "null"]
          description: Set while migrating; it is not yet authoritative.
...
    Project:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
          description: Unique across the collection.
        safeName:
          type: string
        owner:
          type: string
        shortDesc:
          type: string
        longDesc:
          type: string
        group:
          description: An ObjectId, or the populated `Group` on endpoints that populate it.
          oneOf:
            - $ref: "#/components/schemas/ObjectId"
            - $ref: "#/components/schemas/Group"
        path:
          type: string
          description: Datastore-relative path, `/<group.safeName>/<project.safeName>`.
        isPublic:
          type: boolean
          default: false
        accessions:
          type: array
          items:
            type: string
        releaseDate:
          type: string
          description: "ENA submission date, stored as a `DD-MM-YYYY` STRING, not a date."
        doNotSendToEna:
          type: boolean
          default: false
        doNotSendToEnaReason:
          type: string
        nudgeable:
          type: boolean
          default: true
          description: |
            GROUP-DERIVED DEFAULT ON CREATE. `POST /projects/new` honours a
            real JSON boolean from the client and otherwise computes
            `nudgeable = (targetGroup.sendToEna === true)`. A group with
            `sendToEna` absent or `false` therefore produces `nudgeable:
            false`. `PUT /project/toggle-nudgeable` sets the flag afterwards,
            and requires write access to the project's group.

            `komondor-nudge` reads this field directly from MongoDB to decide
            whether to email a project owner. A project where the field is
            ABSENT (predating the field) is skipped and recorded as
            `NUDGEABLE_UNDEFINED` unless
            `NUDGE_INCLUDE_UNDEFINED_NUDGEABLE=true` — so absent and `false`
            are not the same thing to that consumer, even though the API can no
            longer produce an absent value.
        storage:
          $ref: "#/components/schemas/StorageSummary"
        nudges:
          type: array
          description: |
            Timestamps of nudge emails sent. Written by komondor-nudge directly
            against MongoDB (`$push`), never through this API.
          items:
            type: string
            format: date-time
        samples:
          type: array
          description: Virtual. Populated by `GET /project` only.
          items:
            $ref: "#/components/schemas/Sample"
        additionalFiles:
          type: array
          description: Virtual. Populated by `GET /project` only.
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...